Expand SLO workload suite
Move the native query workload under slo-workload, add a shared core harness, and add JDBC and Spring-based workloads so all of them emit one consistent SLO metrics contract.

Rebased onto master after #58 to drop the duplicated JDBC workload commits.
