+++ pom.xml
@@ -30,7 +30,6 @@
         <module>url-shortener-demo</module>
         <module>jdbc</module>
         <module>project-course</module>
-        <module>slo</module>
         <module>slo-workload</module>
     </modules>
