+++ slo-workload/query/src/main/resources/log4j2.xml
