+++ slo-workload/spring-data-jdbc/Dockerfile
@@ -1,23 +1,17 @@
-# Multi-stage Dockerfile for the YDB JDBC SLO workload.
-#
-# The image can be consumed by the YDB SLO action
-# (`ydb-platform/ydb-slo-action`): the workload reads its connection details
-# and run parameters from environment variables and pushes OTLP metrics to the
-# endpoint the action provides.
+# Multi-stage Dockerfile for the Spring Data JDBC SLO workload.
 #
 # Build context: the `ydb-java-examples` repository root. For ydb-jdbc-driver
 # CI the context may also contain `./ydb-jdbc-driver`; when present, the driver
 # is installed from source and the workload is pinned to that exact version.
 #
 # Optional build args:
-#   MAVEN_IMAGE     Builder image. Defaults to `maven:3.9-eclipse-temurin-17`.
-#   RUNTIME_IMAGE   Runtime image. Defaults to `eclipse-temurin:17-jre`.
-#   YDB_JDBC_VERSION Override the ydb-jdbc-driver version under test.
+#   MAVEN_IMAGE       Builder image. Defaults to `maven:3.9-eclipse-temurin-17`.
+#   RUNTIME_IMAGE     Runtime image. Defaults to `eclipse-temurin:17-jre`.
+#   YDB_JDBC_VERSION  Override the ydb-jdbc-driver version under test.
 
 ARG MAVEN_IMAGE=maven:3.9-eclipse-temurin-17
 ARG RUNTIME_IMAGE=eclipse-temurin:17-jre
 
-# ---------- builder ---------------------------------------------------------
 FROM ${MAVEN_IMAGE} AS workload-build
 
 WORKDIR /src
@@ -25,8 +19,6 @@ COPY . /src
 
 ARG YDB_JDBC_VERSION=""
 
-# Install the JDBC driver from source when the driver checkout is present in
-# the build context. Otherwise use YDB_JDBC_VERSION when provided.
 RUN if [ -d /src/ydb-jdbc-driver ]; then \
         cd /src/ydb-jdbc-driver && \
         mvn -B -q \
@@ -52,19 +44,14 @@ RUN if [ -d /src/ydb-jdbc-driver ]; then \
             -DgenerateBackupPoms=false \
             -pl slo-workload ; \
     fi && \
-    mvn -B -q -pl slo-workload/jdbc -am \
+    mvn -B -q -pl slo-workload/spring-data-jdbc -am \
         -DskipTests \
         -Dmaven.javadoc.skip=true \
         package
 
-# ---------- runtime ---------------------------------------------------------
 FROM ${RUNTIME_IMAGE}
 
 WORKDIR /app
+COPY --from=workload-build /src/slo-workload/spring-data-jdbc/target/ydb-slo-spring-data-jdbc-workload.jar /app/ydb-slo-spring-data-jdbc-workload.jar
 
-# The jar's manifest Class-Path points at libs/, so a single `java -jar` call
-# is enough.
-COPY --from=workload-build /src/slo-workload/jdbc/target/ydb-slo-jdbc-workload.jar /app/ydb-slo-jdbc-workload.jar
-COPY --from=workload-build /src/slo-workload/jdbc/target/libs /app/libs
-
-ENTRYPOINT ["java", "-jar", "/app/ydb-slo-jdbc-workload.jar"]
+ENTRYPOINT ["java", "-jar", "/app/ydb-slo-spring-data-jdbc-workload.jar"]
