+++ slo-workload/spring-data-jdbc/src/main/java/tech/ydb/slo/springjdbc/SloExitCodeHolder.java
@@ -1,0 +1,18 @@
+package tech.ydb.slo.springjdbc;
+
+import org.springframework.boot.ExitCodeGenerator;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SloExitCodeHolder implements ExitCodeGenerator {
+    private int exitCode;
+
+    void setExitCode(int exitCode) {
+        this.exitCode = exitCode;
+    }
+
+    @Override
+    public int getExitCode() {
+        return exitCode;
+    }
+}
