+++ slo-workload/spring-data-jpa/src/main/java/tech/ydb/slo/springjpa/SloExitCodeHolder.java
@@ -1,0 +1,18 @@
+package tech.ydb.slo.springjpa;
+
+import org.springframework.boot.ExitCodeGenerator;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SloExitCodeHolder implements ExitCodeGenerator {
+    private int exitCode;
+
+    void setExitCode(int exitCode) {
+        this.exitCode = exitCode;
+    }
+
+    @Override
+    public int getExitCode() {
+        return exitCode;
+    }
+}
