+++ slo-workload/spring-data-jpa/src/main/java/tech/ydb/slo/springjpa/SloSpringDataJpaApplication.java
@@ -1,0 +1,20 @@
+package tech.ydb.slo.springjpa;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+
+/**
+ * Spring Data JPA / Hibernate 6 SLO workload entry point.
+ */
+@SpringBootApplication
+public class SloSpringDataJpaApplication {
+
+    public static void main(String[] args) {
+        int exitCode = SpringApplication.exit(new SpringApplicationBuilder(SloSpringDataJpaApplication.class)
+                .web(WebApplicationType.NONE)
+                .run(args));
+        System.exit(exitCode);
+    }
+}
